#85 chat-modal-window.component replaced by another block inlined in chat.chat-box.component
