--- Cooper/ClientApp/src.v2/app/components/chat-modal-window/chat-modal-window.component.html
@@ -1,35 +1,0 @@
-
-<!-- Modal section -->
-
-  <div class="modal vertically-centered" role="dialog" [style.display]="visibility?'block':'none'">
-  
-  <div class="modal-dialog">
-
-    <!-- Modal content-->
-    <div class="modal-content">
-      
-      <div class="modal-header">
-
-        <button type="button" class="close" data-dismiss="modal" (click)="onCloseClick()">&times;</button>
-
-        <h4 class="modal-title">New Message</h4>
-      </div>
-
-      <!-- model body section -->
-      <div class="modal-body">
-        <p>Send message to: 
-          <input type="text" id="nickname" placeholder="User's nickname" class="input"/></p>
-        <p><textarea matInput id="message" placeholder="Type your message here.." cols="75" rows="5" class="text-area"></textarea></p>
-      </div>
-
-      <div class="modal-footer">
-        <mat-error *ngIf="emptyMessageFieldError">{{emptyMessageErrorContent}}</mat-error>
-        <mat-error *ngIf="emptyNicknameFieldError">{{emptyNicknameErrorContent}}</mat-error>
-        <mat-error *ngIf="notFoundUserError">{{notFoundUserErrorContent}}</mat-error>
-        <button type="button" class="btn btn-default button" data-dismiss="modal" (click)="onSendClick()">Send</button>
-      </div>
-
-    </div>
-
-  </div>
-</div>
