--- Cooper/ClientApp/src.v2/app/components/chat-modal-window/chat-modal-window.component.scss
@@ -1,51 +1,0 @@
-@import "../../constants.scss";
-
-.vertically-centered {
-    top: 50%;
-    position: absolute;
-    margin: 0;
-    -ms-transform: translateY(-50%);
-    transform: translateY(-50%);
-}
-
-
-.text-area {
-    overflow: auto;
-    resize: none;
-    border: 1px solid rgb(184, 184, 184);
-    border-radius: 5px;
-    padding: 5px;
-}
-
-.text-area:focus {
-    border-color: rgb(145, 191, 252);
-    box-shadow: 1px 1px 3px 3px rgb(197, 220, 250);
-}
-
-.input {
-    outline: none;
-	-moz-appearance: none;
-    border: 1px solid rgb(180, 180, 180);
-    border-radius: 5px;
-    padding: 5px;
-}
-
-.input:focus {
-    border-color: rgb(145, 191, 252);
-    box-shadow: 1px 1px 2px 2px rgb(197, 220, 250);
-}
-
-
-.button {
-    outline: none;
-	-moz-appearance: none;
-    border: 1px solid rgb(180, 180, 180);
-    border-radius: 5px;
-}
-
-.button:focus {
-    outline: none;
-	-moz-appearance: none;
-    border-color: rgb(145, 191, 252);
-    box-shadow: 1px 1px 2px 2px rgb(197, 220, 250); 
-}
