+++ Cooper/ClientApp/src.v2/app/components/index.ts
@@ -20,8 +20,6 @@ export { GameListItemComponent } from './game-list-item/game-list-item.component
 export {AvatarCardComponent} from './avatar-card/avatar-card.component';
 export {UploadComponent} from './upload-form/upload-form.component';
 
-export {ChatModalWindowComponent} from './chat-modal-window/chat-modal-window.component';
-
 export {ChatsListComponent} from './chats-list/chats-list.component';
 
 export {ChatBoxComponent} from './chat-box/chat-box.component';
